// src/app/main/apps/procesos/cuadro-horas-30512/util/enum.ts
// export enum TablaCatalogo {
//     //COD_CARRERA_PROGRAMA_ESTUDIOS = 159,
//     //COD_CICLO = 160,
//     COD_CURSO_MODULO = 163,
//     //COD_ESTADO_PLAN_ESTUDIO = 156,
// }

export enum COD_ESTADO_PLAN_ESTUDIO {
    REGISTRADO = 1,
    VIGENTE = 2,
    NO_VIGENTE = 3,
}

export enum FormType {
    REGISTRAR = 1,
    CONSULTAR = 3,
}